feat: move validation utilities for earthquake coordinates; implement geographic bounds checks for Peru
