+++ package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview",
     "deploy": "npm run build && npx gh-pages -d dist",
     "prepare-data": "node scripts/prepare.js",
-    "validate-data": "node src/utils/validateEarthquakeData.js"
+    "validate-data": "node scripts/utils/validateEarthquakeData.js"
   },
   "keywords": [],
   "author": "",
